add separate table_cache
mp works?

mp works

mp works now

mp works now
